pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-tutor-mvp"
version = "0.1.0"
description = "AI Tutor MVP - Personalized Learning Agent"
authors = [
    {name = "AI Tutor Team", email = "ai-tutor@example.com"},
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.0.0",
    "google-generativeai>=0.8.0",
    "google-cloud-aiplatform>=1.0.0",
    "google-cloud-storage>=2.0.0",
    "pydantic>=2.0.0",
    "typing-extensions>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/google/adk-samples"
Repository = "https://github.com/google/adk-samples"
Issues = "https://github.com/google/adk-samples/issues"

[tool.hatch.build.targets.wheel]
packages = ["ai_tutor"]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.ruff]
target-version = "py310"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
